Fixed `goBack` in SiteController to return properly after `authorize` action
Previously `oauth/authorize` returned properly only when user was already logged in.
If user was not logged in it has just redirected to `redirect_uri`.
While it is required to use Oauth2 server `handleAuthorizeRequest`
to return to the given url with oauth2 code and state.

To implement it I added `SaveAuthorizeRequest` (instead of `SetReturnUrl`)
which saves whole authorize request and restores it in `SiteController::goBack()`